requirements.txt
flask
gunicorn
requests
redis
python-dotenv
flask-limiter